requirements.txt
tox
pyyaml
requests
jsonschema
pytest